requirements.txt
pynacl>=1.5.0